backend/requirements.txt
fastapi>=0.110
uvicorn>=0.27
sqlalchemy>=2.0
pydantic>=2.0
qrcode[pil]>=7.4
python-multipart>=0.0.6
pyserial>=3.5
psycopg2-binary>=2.9
python-dotenv>=1.0
pytest>=7.0
httpx>=0.24
